pyproject.toml
[tool.poetry]
name = "maleto"
version = "0.1.5"
description = ""
authors = ["Hadi Zo <hadi.zolfaghaari@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.4.1"
Babel = "^2.9.0"
pymongo = "^3.11.3"
sentry-sdk = "^1.0.0"
prometheus-client = "^0.10.1"
dnspython = "^2.1.0"
ipython = "^7.22.0"
pymongo-inmemory = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
rope = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
maleto = 'maleto:main'